split main up
